requirements.txt
pyqlib>=0.9.6
pandas>=1.5
numpy>=1.23
pyyaml>=6.0
scipy>=1.10
scikit-learn>=1.2
lightgbm>=4.0
matplotlib>=3.7
akshare>=1.16
